Fix the dune dk2nu related issue and use xrootd to read the rootfiles instead

+++ submit_ppfxjobs.py
@@ -7,8 +7,8 @@ import random,time,platform
 ##################AUTHOR AMIT BASHYAL (amitbashyal@gmail.com) ##################
 #USAGE
 #THIS SCRIPT IS TO GENERATE THE PPFX FILES FROM ALREADY EXISTING DK2NU FILES
-# ls path_to_dk2nufiles >file_list.txt
-# python submit_ppfxjobs.py -t file_list.txt -n detector_index
+# ls path_to_dk2nufiles/*.root >file_list.txt
+# python submit_ppfxjobs.py -t file_list.txt -i path/to/input/file/ -o /path/to/outputfile/ -n detector_index
 #END USAGE#####################################################
 
 
@@ -60,10 +60,8 @@ parser.add_option("-o","--output_dir",dest="output_dir",
 parser.add_option("-t","--file_list",dest="file_list",
                   help = "Text file with a list of dk2nu input root files Should be in Same directory as teh job submission script (Only the name of the files) ",default="BLA")
 
-parser.add_option("-i","--input_dir",dest="input_dir",
-                  help="Input Directory where the dk2nu root files live ",default = "$PWD")
 parser.add_option("-n","--detector_index",dest="detector_index",
-                  help="Detector Index for Detector Location ",default = "0")
+                  help="Detector Index for Detector Location ",default = "1")
 
 (options, args) = parser.parse_args()
 print "Output files will be written in ",options.output_dir
@@ -79,10 +77,25 @@ if text_file not in os.listdir(os.getcwd()):
     print text_file, "not in ",os.getcwd()
     sys.exit()
 
+xrootd_input = "root://fndca1.fnal.gov:1094/"
+xrootd_text_file = "xroot_"+text_file
+xrootd_file = open(xrootd_text_file,'w')
+
+with open(text_file,'r') as f:
+    for line in f:
+        if "fndca1.fnal.gov" not in line:
+            __line = xrootd_input+line.replace("/pnfs/","/pnfs/fnal.gov/usr/")
+	    _line = __line.replace("//pnfs","/pnfs")
+	    #print _line
+	    xrootd_file.write(_line)
+	    
+xrootd_file.close()
+#sys.exit()
 _list = []
-_textfile = open(text_file,'r')
+_textfile = open(xrootd_text_file,'r')
 for line in _textfile:
     _list.append(line)
+
 #_list = os.listdir(options.input_dir)
 _pre = []
 index=[]
@@ -101,27 +114,38 @@ for i in range(0,len(_list)):
             if temp[i]!=_pre[i]:
                 index.append(_pre[i])
     _pre = temp
-    #print temp
+    #print _pre
 
-
+#sys.exit()
 __prefix = os.path.basename(_list[0])
 _prefix = __prefix.replace("\n","")
+#print "PREFIX ",_prefix
+#sys.exit()
 if len(index)!=0:
     _temp_prefix = os.path.basename(_list[0])
     temp_prefix = _temp_prefix.replace(index[0],"")
     _prefix = temp_prefix.replace("\n","")
-prefix = _prefix.replace(".root","")
+prefix=""
+if ".dk2nu.root" in _prefix:    
+    prefix = _prefix.replace(".dk2nu.root","")
+else:
+    prefix = _prefix.replace(".root","")
 
-print prefix,_list[0]
-#sys.exit()
+if prefix=="":
+    print "ROOT FILE DOESNT END WITH .dk2nu.root OR .root"
+    sys.exit()
+
+#print prefix,_list[0]
 #sys.exit()
 ntot = len(_list)
+#print prefix,_list,len(_list)
+#sys.exit()
 #now stuff to make the wrapper
 wrapname = "wrapper_doReweight_dk2nu_"+str(int(time.time()))+".sh"
 wrapdir  = CACHE_PNFS_AREA+"wrapdir/"
 wrapfile = wrapdir+wrapname
 temp_text = wrapname.replace(".sh",".txt")
-
+#sys.exit()
 if os.path.exists(wrapfile):
     os.remove(wrapfile)
 if not os.path.exists(wrapdir):
@@ -147,17 +171,19 @@ wrapper.write("source "+ppfx_setup+"\n\n")
 #okay now we create
 #wrapper.write("touch "+temp_text+"\n\n")
 wrapper.write("export run_number=$((${PROCESS}+1)) \n\n")
-wrapper.write('sed "${run_number}q;d" '+text_file+' >'+temp_text+'\n\n')
+wrapper.write('sed "${run_number}q;d" '+xrootd_text_file+' >'+temp_text+'\n\n')
 wrapper.write('file="$(cat '+temp_text+')"'+'\n\n')
-wrapper.write("ifdh cp ${file} "+temp_dir+"\n\n")
-wrapper.write("ls "+temp_dir+"*.root > file_list.txt \n\n")
+#wrapper.write("ifdh cp ${file} "+temp_dir+"\n\n")
+#wrapper.write("ls "+temp_dir+"*.root > file_list.txt \n\n")
 output_filename = prefix+"_${run_number}.root"
-print output_filename
-command = "./bin/doReweight_dk2nu_numix file_list.txt "+temp_dir+output_filename+" scripts/inputs_default.xml "+str(options.detector_index)+"\n\n"
+#print output_filename
+command = "./bin/doReweight_dk2nu_numix "+temp_text+" "+temp_dir+output_filename+" scripts/inputs_default.xml "+str(options.detector_index)+"\n\n"
+print command
+#sys.exit()
 wrapper.write(command)
 
-logfilename = output_filename.replace(".root",".log")
-logfile = logfilename
+logfilename = output_filename.replace(".root","_\$PROCESS.log")
+logfile = os.path.join(str(options.output_dir),logfilename)
 
 #now copy the output file
 wrapper.write("if [ -f "+temp_dir+output_filename+" ];\n")
@@ -165,9 +191,9 @@ wrapper.write("then\n")
 wrapper.write("   ifdh cp -D "+ temp_dir+output_filename+" "+options.output_dir+"\n")
 wrapper.write("fi \n")
 #now copy the log file
-#wrapper.write("if [ -f "+logfile+" ];\n")
+#wrapper.write("if [ -f "+temp_dir+logfile+" ];\n")
 #wrapper.write("then\n")
-#wrapper.write("   ifdh cp -D "+logfile+" "+options.output_dir+"\n")
+#wrapper.write("   ifdh cp -D "+temp_dir+logfile+" "+options.output_dir+"\n")
 #wrapper.write("fi \n")
 wrapper.write("echo 'DONE!!!'\n")
 wrapper.close()
